Add files via upload
A simple C# 6 Net User Control to be used as a Chat box. Where messages appears as bubbles-balloons facebook-like style. "Me" to the right "You" to  the left.

With few lines of bugged code to replace a unicode emoji to an image. ( its commented out. as it replaces only 1 emoji and not always exactly on top)